Dropdown/ComboBox: Fix select first index on open logic
