+++ src/components/Dropdown/V2/Dropdown.reducer.ts
@@ -23,6 +23,7 @@ const reducer = (state = initialState, action: any = {}) => {
       nextState = {
         ...initialItemState,
         isOpen: true,
+        ...payload,
       }
       break
