+++ src/components/Dropdown/V2/Dropdown.types.ts
@@ -44,6 +44,7 @@ export interface DropdownProps extends DropdownMenuDimensions {
   inputValue: string
   isLoading: boolean
   isOpen: boolean
+  isSelectFirstItemOnOpen: boolean
   items: Array<any>
   menuId?: string
   menuOffsetTop: number
